requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorclass==2.2.2
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
docopt==0.6.2
httpie==3.1.0
idna==3.3
multidict==6.0.2
packaging==21.3
Pillow==9.0.1
pip-upgrader==1.4.15
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
terminaltables==3.1.10
urllib3==1.26.9

